+++ lib/output/json.php
@@ -1,0 +1,7 @@
+<?php
+/**
+ * Stub
+ */
+class output_json {
+
+}