+++ lib/output/text.php
@@ -1,0 +1,10 @@
+<?php
+/**
+ * Stub
+ */
+class output_text {
+
+
+
+
+}