+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-# FASE-WP CHANE LOG
+# FASE-WP CHANGE LOG
+
+# V0.12
+- Abstract reporting into classes (with some stubs)
+- Change to autoload to support report classes
+- Added file reporting to HTML output
+- Add options list support (with defaults) to action parser
+- Add do_action line to test1
 
 # V0.11 
 - Rename files and classes.